requirements.txt
requests>=2.31.0
Pillow>=10.0.0
img2pdf>=0.4.4
tk